Update locale management
